refactor: split runtime and surface modal blocked sessions
